web/requirements.txt
asgiref==3.4.1
Django==3.2.7
pytz==2021.1
sqlparse==0.4.2
psycopg2-binary==2.9.1

mozilla-django-oidc
django-mirage-field
